Makefile.am: remove generated documentation on make distclean
A rule to remove the entire doc/ tree on make clean was added
in polylib-5.22.3-55-gdba808d (Configure and makefile.am's
compatibility with bsd-make, Wed Feb 24 14:37:21 2010 +0100).
This was removed in polylib-5.22.5-1-g52d9e6d
(Mon Apr 12 16:59:36 2010 +0200), possibly because it removed
too much.

However, the generated documentation should be removed on make distclean
for make distcheck to succeed.